Update patch set 1
Patch Set 1:

Does anyone else have any opinions on this?  As noted in the comments, project hooks are no longer restored if the user modifies them.  This is a change in behavior but I think it makes sense.

If people would prefer that behavior to not change I can adjust this so that it we still delete all project hooks (as opposed to deleting *all* hooks) which would cause them to be fully restored by _InitHooks.

Patch-set: 1