Update patch set 1
Patch Set 1:

Gerrit is acting funny and I'm not able to add davidjames@google.com, but this message is for him when I get in touch with him:

This is related to:
I5e2f4e1a7abb56f9d3f310fa6fd0c17019330ecd

The question is, why does .project-objects use project names instead of the the path like .projects?  If there's not a reason, the right fix here seems to be to move these over to a consistent naming pattern.

Patch-set: 1