Update patch set 1
Patch Set 1:

The straightforward way of correcting the issue risks losing locally created refs, something like:

1) Create a new directory in project-objects for the new project as project-objects/<name>
2) Change the links from the projects/<path> directory to point at the new project-objects dir. Most relevantly the 'objects' link.
3) Copy the unlinked pieces of the new project-objects directory to the projects directory? This overwrites refs and packed refs, so some local refs from the old tree may be lost.
4) Update <path>/.git with new links and run checkout.

It may be possible to re-work the structure of the projects/ directory to allow saving old versions of the same path that use different projects (maybe by adding the project name to the path?). Even, then the user experience doesn't seem great if some of the refs they created disappeared after a re-init and sync. Following this path would also mean that the new version of repo would no longer be compatible with the directory structure created by older versions.

Combining the refs and objects from both projects runs the risk of ref collisions, especially since many projects use the same branch names.

If you are ok removing local refs during a sync after a re-init, we can recover in a manner similar to the one in the --force-sync follow on change without adding the flag. Should we follow that approach?

Patch-set: 1
Label: Verified=0