Update patch set 1
Patch Set 1:

(1 comment)

OK, if you are just trying to make the failure more obvious, that's fine.

So it looks like your solution is to pull _ReferenceGitDir out of the conditionals it gets used in so it's run every time we want to init anything...is this a hacky way of accomplishing the goal?  Is there not a simple test we could write that wouldn't break up each of conditionals?

Does that make sense?

Patch-set: 1