Print output of stderr in realtime when doing 'git fetch'
When doing a 'git fetch' the output of stderr would not be printed until
after the entire command was executed.  This was making it difficult to
synchronize how the stdout and stderr were related to each other.  This
patch makes it so that the stderr output is printed out in realtime and
should be relatively in sync with the stdout output.

Updated GitCommand to have a 'tee_stderr' option.  If this is selected,
then the stderr output will be printed in realtime and it will also be
saved as in the normal usage of 'capture_stderr'.

Change-Id: Ifdae138e008f80a59195f9f43c911a1a5210ec60