Update patch set 7
Patch Set 7:

> Hm, fcntl isn't available on Windows AFAICT. But I honestly don't know if we support repo on Windows.

Last time I contributed non-windows code, it was (paraphrased) "we don't support it, but people want to, so avoid it".


Personally, I would use a Queue for all output. The code would be much simpler, be compatible with Windows, and, at a later stage, could be used for anything that outputs (can't remember what does off the top of my head).

Patch-set: 7
Label: Verified=0