Update patch set 8
Patch Set 8: Code-Review+1

I went with the select method instead of the StreamWatcher as it was easy for me to finish it.  I did try the StreamWatcher method but was having some difficulties and the select method was working.

Patch-set: 8
Reviewer: Gerrit User 1017359 <1017359@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+1