fetch commit-msg hook from gerrit
This will fetch the latest hook, which removes the need to update the
hook anymore.

The hook is fetched on a per project basis, allowing for the hook to
be fetched from different teams' gerrit instances.
If a review is not set, fall back to fetching the hook from
gerrit-review.googlesource.com.

A simple file structure along with mtime checking is used to ensure
the hook is only downloaded when needed.

If there is an issue fetching the hook, fall back silently to a
backup hook.

Change-Id: Ib3363d1205b9891e1fc6d87dcceab47d4b7e3825