Update patch set 1
Patch Set 1: Code-Review-1

Wanted to get your opinion on this before I go any further with it.

The intention is to have ALL the output go through logging, and at a later point, add in methods to register file like objects (subprocess pipes) to be watched, and sent to Logger in real time.

So the short term aim is get rid of directly using stdout/stderr/print, and long term is making it handle pipes.

Patch-set: 1
Label: Code-Review=-1