Override https_proxy only for darwin
repo and main.py configured urllib2.request.ProxyHandler to use the
value of os.environ['http_proxy'] for both http and https. This broke
shops that segregate http and https traffic to different proxies as
setting https_proxy had effect on tools like curl but didn't work for
repo.

urllib2 (and urllib in Python3) has built-in support for handling the
variables correctly, and if they are not set use OS specific mechanisms
to obtain proxies. So overriding this in ProxyHandler should not be
necessary.

However, git_command.py and project.py has explicit handling of
http_proxy if sys.platform == 'darwin'. Assuming that this is needed to
work around some specific bug, I leave the ProxyHandler configuration in
place for darwin so that existing behavior is retained on that platform.

Change-Id: I302aefa17f7ae5d066761b410a4eaf17898222ea