Update patch set 4
Patch Set 4:

The option is safe to use for versions that don't support the file protocol, though a warning will be posted: "Warning: unrecognized protocol 'file'" (silent doesn't hide it either).

To detect it, you need to check curl --version, which *should* always return it in the form "Protocols: dict file..."

I would be tempted to actually use httplib and Cookie for this, since curl isn't usually on Windows. I'll look in to writing it up in a bit

Patch-set: 4