RFC: Add --prune option to fetch when syncing a mirror repo
When syncing a mirror repo, add the --prune option to the fetch
command to force removal of stale refs from the mirror.

RFC because I'm not sure if this is something that should be done
by default, or made configurable.  Or maybe it's better to not do
this at all and the user should set "fetch.prune" to true in their
global git config file?

Change-Id: I4b43b2a5c86b9915627887c16f6569066f3ab978