Resolve remote aliases when saving manifests
Projects should always refer to the remote names in manifest files.  The
XmlManifest.Save function failed to do this when default remote had an
alias, and probably in other cases too.

The safest way to find the proper remote element name is to compare the
actual fetch URLs.  Also check the names and aliases of matching remotes
in case there are several remote elements with the same URL.

Change-Id: I8a020e6ebb3600fe47f819cd578a9acd3ef235c9