Update patch set 3
Patch Set 3:

The last version I uploaded explains the exact command to use. However, the explanation isn't accurate if one is running repo from a subdirectory. I can revise the patch if you provide some guidance about how to get that information programmatically (linking me to an existing example in the codebase would be enough).

For users that have to force-sync all of their trees, I hope that at some point they would see the pattern and decide to do all of their trees at once (after double checking that it's safe to do so).

Patch-set: 3