GITC: Add repo start support.
Add repo start support for GITC checkouts. If the user is in
the GITC FS view, they can now run repo start to check out
the sources and create a new working branch. The GITC manifest
is updated so that the GITC filesystem knows to display the
local copy of the sources when being viewed.

Change-Id: I7e4809d1c4fc43c69b26f2f1bebe45aab0cae628