Fix gitc-init behavior
With gitc-init, a gitc client may be specified using '-c'. If we're not
currently in that client, we need to change directories so that we don't
affect the local checkout, and to ensure that repo is checked out in the
new client.

This also makes '-c' optional if already in a gitc client, to match the
rest of the init options.

Change-Id: Ib514ad9fd101698060ae89bb035499800897e9bd