Update patch set 1
Patch Set 1:

This should fix 24105716, since we weren't going through repo:_Init if there was a local copy of repo.

1) There's a check in gitc_init.py to ensure that the '-c' matches the local directory, that should handle all the currently working usecases. If you're in a different checkout, it will still behave like 24105716. I do need to add a check so that old 'repo', in a non-gitc client, and no '-c' prints an error instead of throwing an exception.

I'll upload a patch to improve the wording so that it points them to go/gitc to make sure they update repo (we can remove that eventually). If this goes in before I do the next repo release, they'll get a separate notification to update repo wrapper, since the version has been updated since the last release.

2) A new repo wrapper will always move to the gitc client before doing anything else, so the local version doesn't matter. If there's an older version in the gitc client, you just don't get the optional '-c' behavior until it updates.

Patch-set: 1
Label: Verified=0