Do not require force sync for unchanged projects
Many third party Android teams switch the upstream source for projects
frequently, particularly shortly after a new major version is released.
For example, they may switch a project from AOSP to CAF for hardware
changes or from AOSP to GitHub for local changes.  This is typically
done with several dozen projects.  Most developers on a team won't have
any local changes or branches for most of these projects.  So it is a
needless burden to require a force sync in these cases.

Change-Id: I79eb6586ce38c358d17fa10f1ce0b5f37a567bf5