Add option to rebase onto project's manifest version
Some teams have a continuous build server that would mark certain manifest green
and safe to sync to.  Then team members could repo sync to that particular manifest
file and make sure they always sync to a green build.  But if she/he has some local
changes and wants to rebase, currently it would be a manual process to find the
correct version to rebase onto.  This patch helps with that use case by automate the
process to rebase onto the currently synced manifest version.

Change-Id: I847c9eb6addf7f84fd3f5594fbf8c0bcc103f9a5