repo: Add check of REPO_URL env variable
We want to be able to run repo on a system that is not connected to the
Internet and cannot access https://gerrit.googlesource.com. We can
put a clone of that repos there, but would prefer to use the stable
version of the repo script instead of a locally modified version.

This commit adds a check for the REPOS_URL environment variable. If that
is set and not empty its value will be set in the REPOS_URL global in repo.
Otherwise the standard path will be used.

Change-Id: I0616f5f81ef75f3463b73623b892cb5eed6bb7ba