Update patch set 1
Patch Set 1:

> I noticed when going through this exercise that there is a
 > hooks/commit-msg file in the repository, but it's slightly
 > different than the hook installed by the curl command specified in
 > the SUBMITTING_PATCHES files. Should they be the same? I used the
 > curl version.

The commit-msg hook is maintained in the Gerrit project, and the version you get via curl will depend on what version of Gerrit the server is running.

When the hook is installed via repo it'll use the one in the repo git.

The version from gerrit is usually added into git-repo after each gerrit release, to keep them in sync.  I'll check if that's been done recently.

Patch-set: 1