Fix symlinking of new projects
We weren't copying these list, and the += was actually changing the
underlying lists.

So when a new project was added to the manifest, we run
_CheckDirReference against the manifest project with share_refs=True,
which added the working_tree_* to the shareable_* lists. Then, when we
loaded the new manifest and create the new project, it uses the lists
that already contains the working_tree_* files, even though we passed
share_refs=False.

This happens reliably under the above conditions, but doesn't seem to
happen when syncing a fresh tree. So we've got a mixture of links that
may need to be cleaned up later. This patch will just stop it from
happening in the future.

Change-Id: Ib7935bfad78af1e494a75e55134ec829f13c2a41