Update patch set 3
Patch Set 3:

Actually, is there ever a valid URL that doesn't start with the form:

xxxx://

Specifically that "//"

The script currently requires that it start with "http", "https" or "sso". The first two require "//", my question was about "sso". I assume it stands for single sign on, but I have never seen a URI actually begin with "sso".

Am I fine assuming that all urls should contain "//", because currently if they don't this will mangle the host and thus give a:

"urlopen error no host given" error when trying to connect (A reasonable error)

Patch-set: 3