Better support for https-adding proxies
The local proxy we use behind persistent-https always uses http for the
local connection, and sets up the appropriate SSL connections itself. It
used to rewrite the redirect urls from https to http, but is no longer
doing so. Instead, use a custom HTTPRedirectHandler so that redirection
to an https site still works.

Also, turn on HTTP debugging when --trace is set so that issues are
easier to debug in the future.

Change-Id: I5bb21a8a64142dea4a827faedf4255fb867ecd51