Update patch set 3
Patch Set 3:

For each individual fetch, we already retry once on failure, and wait between 30 and 45 seconds (randomized) between tries. Is this not enough for you, or are you hitting an error in some other path not covered by this retry?

https://gerrit.googlesource.com/git-repo/+/master/project.py#1963

I see this commit was originally created before this was implemented in repo (~Oct 2011)

Patch-set: 3