Add a superproject metarepository for atomic syncs
Gerrit learned to cope with cross repository atomic submissions over the
course of the last year. While this helps users of Gerrit, it is only half
the solution. We need to find a way to transport this atomicity from
the server side (Gerrit) to the users.

For that Gerrit also strengthened its superproject subscription feature.
That allows a project to subscribe to other projects that are submodules
and whenever one of these submodule projects is updated the gitlink in
the subscribing project is updated as well. With those two Gerrit features
in place it is possible to craft a superproject of larger software
ecosystems, e.g. Android tracked in one single repository.

While the long term plan is to improve Git submodule handling to cope with
such a single tracking repository containing a lot of submodules, we also
want to see impact in the short term, which is what this is change is
supposed to deliver.

This change adds the key "default.use_superproject" to the manifest format
that is an url which points to the superproject. If configured the
superproject is then cloned as a MetaProject into .repo/revisions[.git].

If a project revision is not set in the manifest it is filled from the
revisions superproject.

This allows for minimal modification of the manifest by just adding
the default.use_superproject key and enjoy atomic syncs by using the
superproject.

Change-Id: I6c19a88a98518c9a8b45d46aae9714f9b4143ba8