Update patch set 1
Patch Set 1:

> There's already two implementations of basically this functionality
 > in the code already. They just trigger in different circumstances
 > :)
 > 
 > The --optimized_fetch option to sync will skip the network sync
 > (with need_to_fetch in project.py) if the revisionExpr is a SHA1.
 > 
 > If -c (current branch only) is set, and revisionExpr is a SHA1, we
 > exit from _RemoteFetch early.
 > 
 > These should probably be updated to use revisionId (and make sure
 > everything else mixes revisionExpr and revisionId properly)

IIUC these 2 optimizations only care about corner cases?
When revisionExpr is a sha1 (that is not close to always, but rather a "sometimes") we omit fetching if we have that sha1 already.

the optimized_fetch only fetches one branch instead of many. But the fetching itself has some overhead (like connecting to the server, waiting for the server to compute the pack etc)

Once the prior commit is rolled out and used, a daily "repo sync -Q" of AOSP would result in about 10-20 fetches and 500 skips.

I'll look into the other options and this and how they can be used for different corner cases.

Patch-set: 1
Reviewer: Gerrit User 1017837 <1017837@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0