Update patch set 1
Patch Set 1:

> Interesting. It sounds like both usecases should be using `git
 > rev-parse --verify ...^{commit}`. And yes, dealing with the
 > revisionExpr vs revisionId vs passed in parameter needs to be
 > cleaned up through most of the project class

I overlooked the ^0 in _CheckForSha1, so it actually checks for the 0th parent of the commit (i.e. the commit itself). But to compute that, Git parses the commit, which is triggering the existence check.

So _CheckForSha1 is not broken. I will use that instead of adding a new function. Thanks!

Patch-set: 1